sql/sqlspec: export mightheredoc helper
